Moved Habit displaying and updating functionality to HabitTableRows
